refactor: etf 단건 조회 inav 필드 double로 수정

+++ frontend/WYE/app/src/main/java/com/d102/wye/data/remote/dto/response/EtfDetailResponse.kt
@@ -19,8 +19,8 @@ data class EtfDetailResponse(
     @SerializedName("roe") val roe: Double,
     @SerializedName("aum") val aum: Long,
     @SerializedName("listingDate") val listingDate: String,
-    @SerializedName("inav") val inav: Long,
-    @SerializedName("inavChangeAmount") val inavChangeAmount: Long,
+    @SerializedName("inav") val inav: Double,
+    @SerializedName("inavChangeAmount") val inavChangeAmount: Double,
     @SerializedName("inavChangeRate") val inavChangeRate: Double,
 )
